Pequenos ajustes para acomodar arquitec. hexagonal
