+++ src/main/java/com/example/hito1_ad_crud/Hito1AdCrudApplication.java
@@ -1,10 +1,8 @@
 package com.example.hito1_ad_crud;
 
-import com.example.hito1_ad_crud.connection.MyConnection;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.stereotype.Component;
 
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 public class Hito1AdCrudApplication {
