+++ src/main/java/com/example/hito1_ad_crud/domain/Libro.java
@@ -1,20 +1,23 @@
-package com.example.hito1_ad_crud.entity;
+package com.example.hito1_ad_crud.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
-
+@Component
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Component
-public class User {
+public class Libro {
 
-    private int idUser;
+    private int idLibro;
     private String name;
-    private String nif;
+    private String author;
+    private String editorial;
+    private int num_pages;
+    private boolean disponible;
+    private int idUser;
 
 }
