--- src/main/java/com/example/hito1_ad_crud/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.hito1_ad_crud.repository;
-
-import com.example.hito1_ad_crud.entity.User;
-
-@org.springframework.stereotype.Repository
-public interface UserRepository extends com.example.hito1_ad_crud.repository.Repository<User, Integer> {
-
-
-}
