+++ src/main/java/com/example/hito1_ad_crud/service/Repository.java
@@ -1,4 +1,4 @@
-package com.example.hito1_ad_crud.repository;
+package com.example.hito1_ad_crud.service;
 
 import java.util.List;
