+++ src/main/java/com/example/hito1_ad_crud/service/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.hito1_ad_crud.service;
+
+import com.example.hito1_ad_crud.domain.User;
+
+@org.springframework.stereotype.Repository
+public interface UserRepository extends com.example.hito1_ad_crud.service.Repository<User, Integer> {
+
+
+}
